@@ -3,13 +3,13 @@ import org.apache.tools.ant.taskdefs.condition.Os
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.2'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         applicationId "com.nowsecure.android.vts"
         minSdkVersion 15
-        targetSdkVersion 25
+        targetSdkVersion 28
         versionCode 13
         versionName "v.13"
     }
@@ -41,19 +41,19 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile 'com.android.support:cardview-v7:25.3.1'
-    compile 'com.android.support:design:25.3.1'
-    compile 'com.evernote:android-job:1.1.8'
-    compile 'com.github.paolorotolo:appintro:4.1.0'
-    compile 'com.squareup.retrofit2:adapter-rxjava:2.2.0'
-    compile 'com.squareup.retrofit2:converter-gson:2.2.0'
-    compile 'com.squareup.retrofit2:retrofit:2.2.0'
-    compile 'commons-io:commons-io:2.5'
-    compile 'io.reactivex:rxandroid:1.2.1'
-    compile 'org.apache.commons:commons-compress:1.13'
-    compile 'com.afollestad.material-dialogs:core:0.9.4.2'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'com.evernote:android-job:1.2.6'
+    implementation 'com.github.paolorotolo:appintro:v5.1.0'
+    implementation 'com.squareup.retrofit2:adapter-rxjava:2.5.0'
+    implementation 'com.squareup.retrofit2:converter-gson:2.5.0'
+    implementation 'com.squareup.retrofit2:retrofit:2.5.0'
+    implementation 'commons-io:commons-io:2.6'
+    implementation 'io.reactivex:rxandroid:1.2.1'
+    implementation 'org.apache.commons:commons-compress:1.18'
+    implementation 'com.afollestad.material-dialogs:core:0.9.6.0'
 }
 
 def ndkDir = project.android.ndkDirectory
@@ -97,15 +97,4 @@ clean.dependsOn 'cleanNative'
 
 tasks.withType(JavaCompile) {
    compileTask -> compileTask.dependsOn buildNative
-}
-
-configurations.all {
-    resolutionStrategy.eachDependency { DependencyResolveDetails details ->
-        def requested = details.requested
-        if (requested.group == 'com.android.support') {
-            if (!requested.name.startsWith("multidex")) {
-                details.useVersion '25.3.1'
-            }
-        }
-    }
 }
