@@ -1,26 +1,19 @@
 package fuzion24.device.vulnerability.test.adapter;
 
 import android.content.Context;
-import android.graphics.Color;
-import android.support.v7.widget.RecyclerView;
-import android.text.TextUtils;
 import android.util.Log;
 import android.util.Pair;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.nowsecure.android.vts.R;
 
-import org.json.JSONArray;
-import org.json.JSONObject;
-
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -28,7 +21,6 @@ import java.util.Map;
 import fuzion24.device.vulnerability.test.VulnerabilityDescriptor;
 import fuzion24.device.vulnerability.test.VulnerabilityTestResult;
 import fuzion24.device.vulnerability.test.adapter.viewholder.RecyclerItemViewHolder;
-import fuzion24.device.vulnerability.vulnerabilities.helper.BinaryAssets;
 
 public class RecyclerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
 
