@@ -12,6 +12,8 @@ import com.nowsecure.android.vts.R;
 
 import fuzion24.device.vulnerability.util.SharedPreferencesUtils;
 
+import androidx.annotation.NonNull;
+
 public class AppIntroActivity extends AppIntro {
 
     @Override
@@ -38,19 +40,19 @@ public class AppIntroActivity extends AppIntro {
             .negativeText(R.string.global_notsharing)
             .onPositive(new MaterialDialog.SingleButtonCallback() {
                 @Override
-                public void onClick(MaterialDialog materialDialog, DialogAction dialogAction) {
+                public void onClick(@NonNull MaterialDialog materialDialog, @NonNull DialogAction dialogAction) {
                     SharedPreferencesUtils.setIsAutomaticSharingEnable(AppIntroActivity.this, true);
                 }
             })
             .onNegative(new MaterialDialog.SingleButtonCallback() {
                 @Override
-                public void onClick(MaterialDialog materialDialog, DialogAction dialogAction) {
+                public void onClick(@NonNull MaterialDialog materialDialog, @NonNull DialogAction dialogAction) {
                     SharedPreferencesUtils.setIsAutomaticSharingEnable(AppIntroActivity.this, false);
                 }
             })
             .onAny(new MaterialDialog.SingleButtonCallback() {
                 @Override
-                public void onClick(MaterialDialog materialDialog, DialogAction dialogAction) {
+                public void onClick(@NonNull MaterialDialog materialDialog, @NonNull DialogAction dialogAction) {
                     SharedPreferencesUtils.disableFirstRun(AppIntroActivity.this);
 
                     Intent mainIntent = new Intent(AppIntroActivity.this, MainActivity.class);
