@@ -7,20 +7,22 @@ import android.content.res.Configuration;
 import android.graphics.Color;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.coordinatorlayout.widget.CoordinatorLayout;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.snackbar.Snackbar;
 
 import com.afollestad.materialdialogs.DialogAction;
 import com.afollestad.materialdialogs.MaterialDialog;
@@ -40,7 +42,6 @@ import fuzion24.device.vulnerability.test.ShareViaGoogleForm;
 import fuzion24.device.vulnerability.test.VulnerabilityTestResult;
 import fuzion24.device.vulnerability.test.VulnerabilityTestRunner;
 import fuzion24.device.vulnerability.test.adapter.RecyclerAdapter;
-import fuzion24.device.vulnerability.test.job.AndroidVTSJobCreator;
 import fuzion24.device.vulnerability.test.job.CheckForUpdatesJob;
 import fuzion24.device.vulnerability.util.DeviceInfo;
 import fuzion24.device.vulnerability.util.SharedPreferencesUtils;
@@ -66,7 +67,7 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        Toolbar toolbar = findViewById(R.id.toolbar);
         toolbar.setTitleTextColor(Color.WHITE);
         setSupportActionBar(toolbar);
         getSupportActionBar().setTitle(R.string.app_name);
@@ -77,23 +78,23 @@ public class MainActivity extends AppCompatActivity {
             testResults = new ArrayList<>();
         }
 
-        RecyclerView recyclerView = (RecyclerView) findViewById(R.id.recyclerView);
-        coordinatorLayout = (CoordinatorLayout) findViewById(R.id.coordinatorLayout);
-        emptyView = (TextView) findViewById(R.id.emptyView);
+        RecyclerView recyclerView = findViewById(R.id.recyclerView);
+        coordinatorLayout = findViewById(R.id.coordinatorLayout);
+        emptyView = findViewById(R.id.emptyView);
         recyclerAdapter = new RecyclerAdapter(MainActivity.this, testResults);
 
         recyclerView.setLayoutManager(new LinearLayoutManager(MainActivity.this));
         recyclerView.setAdapter(recyclerAdapter);
 
-        final TextView tvKernelVersion = (TextView) findViewById(R.id.kernelVersion);
-        final TextView tvBuildFingerPrint = (TextView) findViewById(R.id.buildFingerPrint);
-        final TextView tvBuildID = (TextView) findViewById(R.id.buildID);
-        final TextView tvBuildBrand = (TextView) findViewById(R.id.buildBrand);
-        final TextView tvBuildManuf = (TextView) findViewById(R.id.buildManufacturer);
-        final TextView tvBuildModel = (TextView) findViewById(R.id.buildModel);
-        final TextView tvBuildRelease = (TextView) findViewById(R.id.buildRelease);
-        final TextView tvBuildSDK = (TextView) findViewById(R.id.buildSDK);
-        final TextView tvBuildABIList = (TextView) findViewById(R.id.buildABIList);
+        final TextView tvKernelVersion = findViewById(R.id.kernelVersion);
+        final TextView tvBuildFingerPrint = findViewById(R.id.buildFingerPrint);
+        final TextView tvBuildID = findViewById(R.id.buildID);
+        final TextView tvBuildBrand = findViewById(R.id.buildBrand);
+        final TextView tvBuildManuf = findViewById(R.id.buildManufacturer);
+        final TextView tvBuildModel = findViewById(R.id.buildModel);
+        final TextView tvBuildRelease = findViewById(R.id.buildRelease);
+        final TextView tvBuildSDK = findViewById(R.id.buildSDK);
+        final TextView tvBuildABIList = findViewById(R.id.buildABIList);
 
         devInfo = DeviceInfo.getDeviceInfo();
 
@@ -115,7 +116,7 @@ public class MainActivity extends AppCompatActivity {
 
         tvBuildABIList.setText(sb.toString());
 
-        FloatingActionButton fabStart = (FloatingActionButton) findViewById(R.id.fabStart);
+        FloatingActionButton fabStart = findViewById(R.id.fabStart);
 
         fabStart.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -195,19 +196,19 @@ public class MainActivity extends AppCompatActivity {
                     .negativeText(R.string.global_notsharing)
                     .onPositive(new MaterialDialog.SingleButtonCallback() {
                         @Override
-                        public void onClick(MaterialDialog materialDialog, DialogAction dialogAction) {
+                        public void onClick(@NonNull MaterialDialog materialDialog, @NonNull DialogAction dialogAction) {
                             SharedPreferencesUtils.setIsAutomaticSharingEnable(MainActivity.this, true);
                         }
                     })
                     .onNegative(new MaterialDialog.SingleButtonCallback() {
                         @Override
-                        public void onClick(MaterialDialog materialDialog, DialogAction dialogAction) {
+                        public void onClick(@NonNull MaterialDialog materialDialog, @NonNull DialogAction dialogAction) {
                             SharedPreferencesUtils.setIsAutomaticSharingEnable(MainActivity.this, false);
                         }
                     })
                     .onAny(new MaterialDialog.SingleButtonCallback() {
                         @Override
-                        public void onClick(MaterialDialog materialDialog, DialogAction dialogAction) {
+                        public void onClick(@NonNull MaterialDialog materialDialog, @NonNull DialogAction dialogAction) {
                             invalidateOptionsMenu();
                         }
                     })
@@ -273,7 +274,7 @@ public class MainActivity extends AppCompatActivity {
             .positiveText(R.string.dismiss)
             .onAny(new MaterialDialog.SingleButtonCallback() {
                 @Override
-                public void onClick(MaterialDialog materialDialog, DialogAction dialogAction) {
+                public void onClick(@NonNull MaterialDialog materialDialog, @NonNull DialogAction dialogAction) {
                     if (SharedPreferencesUtils.isAutomaticSharingEnable(MainActivity.this)) {
                         try {
                             JSONObject json = VulnerabilityResultSerialzier.serializeResultsToJson(testResults, devInfo);
@@ -293,7 +294,7 @@ public class MainActivity extends AppCompatActivity {
             dialogBuilder.negativeText(R.string.my_device_is_vulnerable_info)
                 .onNegative(new MaterialDialog.SingleButtonCallback() {
                     @Override
-                    public void onClick(MaterialDialog materialDialog, DialogAction dialogAction) {
+                    public void onClick(@NonNull MaterialDialog materialDialog, @NonNull DialogAction dialogAction) {
                         Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse("https://www.nowsecure.com/blog/2015/11/18/my-device-is-vulnerable-now-what/"));
                         startActivity(browserIntent);
                     }
@@ -302,7 +303,7 @@ public class MainActivity extends AppCompatActivity {
 
         View view = dialogBuilder.show().getCustomView();
 
-        TextView state = (TextView) view.findViewById(R.id.scan_details_device_state);
+        TextView state = view.findViewById(R.id.scan_details_device_state);
         if (numberOfFailed > 0) {
             state.setText(R.string.scan_details_vulnerable_message);
             state.setTextColor(getResources().getColor(R.color.red));
@@ -311,10 +312,10 @@ public class MainActivity extends AppCompatActivity {
             state.setTextColor(getResources().getColor(R.color.green));
         }
 
-        TextView numberOfTests = (TextView) view.findViewById(R.id.scan_details_number_tests);
+        TextView numberOfTests = view.findViewById(R.id.scan_details_number_tests);
         numberOfTests.setText(String.valueOf(results.size()));
 
-        TextView numberOfFailedTests = (TextView) view.findViewById(R.id.scan_details_failed_tests);
+        TextView numberOfFailedTests = view.findViewById(R.id.scan_details_failed_tests);
         numberOfFailedTests.setText(String.valueOf(numberOfFailed));
     }
 
@@ -323,7 +324,6 @@ public class MainActivity extends AppCompatActivity {
             .setPeriodic(TimeUnit.DAYS.toMillis(7))
             .setRequirementsEnforced(true)
             .setRequiredNetworkType(JobRequest.NetworkType.CONNECTED)
-            .setPersisted(true)
             .build()
             .schedule();
     }
@@ -340,7 +340,7 @@ public class MainActivity extends AppCompatActivity {
     }
 
     @Override
-    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
         switch (requestCode) {
             case CREATE_CHECK_FOR_UPDATES_JOB:
                 if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
@@ -357,12 +357,8 @@ public class MainActivity extends AppCompatActivity {
     }
 
     @Override
-    public void onConfigurationChanged(Configuration newConfig) {
-        super.onConfigurationChanged(newConfig);
-    }
-
-    @Override
     protected void onSaveInstanceState(final Bundle outState) {
+        super.onSaveInstanceState(outState);
         outState.putSerializable(SERIALIZABLE_RESULTS, testResults);
     }
 
